1.0.1 update

+++ TechnicallyAScannerBuff/Main.cs
@@ -3,6 +3,7 @@ using RoR2;
 using R2API;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
+using BepInEx.Configuration;
 
 namespace TimedSeeSChest
 {
@@ -10,16 +11,20 @@ namespace TimedSeeSChest
     [BepInDependency(PrefabAPI.PluginGUID)]
 
     // Metadata
-    [BepInPlugin("Samuel17.TechnicallyAScannerBuff", "TechnicallyAScannerBuff", "1.0.0")]
+    [BepInPlugin("Samuel17.TechnicallyAScannerBuff", "TechnicallyAScannerBuff", "1.0.1")]
 
     public class Main : BaseUnityPlugin
     {
         // Fields
-        public GameObject newCloakedChestPrefab;
-        public InteractableSpawnCard iscChest2Stealthed;
+        public static GameObject newCloakedChestPrefab;
+        public static InteractableSpawnCard iscChest2Stealthed;
+        public static bool scannerUsedThisStage = false;
+        public static SceneDirector currentSceneDirector;
 
         // Config fields
-        public static int chestCount = 1;
+        public static ConfigEntry<int> chestCount { get; private set; }
+        public static ConfigEntry<bool> largeRarity { get; private set; }
+        public static ConfigEntry<bool> scannerRequired { get; private set; }
 
         // Load addressables
         public static GameObject cloakedChestPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/Base/Chest1StealthedVariant/Chest1StealthedVariant.prefab").WaitForCompletion();
@@ -32,13 +37,16 @@ namespace TimedSeeSChest
             Log.Init(Logger);
 
             // Configs
-            chestCount = Config.Bind("Cloaked Chests", "Chest Count", 1, "The amount of guaranteed Cloaked Chests per stage.").Value;
+            chestCount = Config.Bind("Cloaked Chests", "Chest Count", 1, "The amount of guaranteed Cloaked Chests per stage.");
+            largeRarity = Config.Bind("Cloaked Chests", "Improved Item Rarity", true, "Item rarity is equal to a Large Chest. False makes it equal to a regular Chest.");
+            scannerRequired = Config.Bind("Cloaked Chests", "Scanner is Required", false, "If set to true, the guaranteed Cloaked Chests can only be found by activating the Radar Scanner at least once.");
 
             // Create new Cloaked Chest type
             SetupNewCloakedChest();
 
             // Populate stages with it
-            On.RoR2.SceneDirector.PopulateScene += SpawnCloakedChest;
+            SceneDirector.onPostPopulateSceneServer += StageStart;
+            if (scannerRequired.Value) { On.RoR2.EquipmentSlot.FireScanner += SpawnCloakedChestWithRadar; } 
         }
 
         private void SetupNewCloakedChest()
@@ -47,7 +55,10 @@ namespace TimedSeeSChest
             newCloakedChestPrefab = PrefabAPI.InstantiateClone(cloakedChestPrefab, "Chest2StealthedVariant");
 
             // Set its loot table to match a Large Chest
-            newCloakedChestPrefab.GetComponent<ChestBehavior>().dropTable = largeChestDropTable;
+            if (largeRarity.Value)
+            {
+                newCloakedChestPrefab.GetComponent<ChestBehavior>().dropTable = largeChestDropTable;
+            }
 
             // Prevent it from being locked during the Teleporter event. Yes we're gonna need a wholeass hook for this.
             newCloakedChestPrefab.GetComponent<PurchaseInteraction>().setUnavailableOnTeleporterActivated = false;
@@ -59,18 +70,37 @@ namespace TimedSeeSChest
             iscChest2Stealthed.prefab = newCloakedChestPrefab;
         }
 
-        private void SpawnCloakedChest(On.RoR2.SceneDirector.orig_PopulateScene orig, SceneDirector self)
+        private void StageStart(SceneDirector sceneDirector)
         {
-            orig(self);
+            currentSceneDirector = sceneDirector;
+            scannerUsedThisStage = false;
 
+            if (!scannerRequired.Value)
+            {
+                SpawnCloakedChest(sceneDirector);
+            }
+        }
+
+        private bool SpawnCloakedChestWithRadar(On.RoR2.EquipmentSlot.orig_FireScanner orig, EquipmentSlot self)
+        {
+            if (currentSceneDirector && !scannerUsedThisStage)
+            {
+                scannerUsedThisStage = true;
+                SpawnCloakedChest(currentSceneDirector);
+            }
+            return orig(self);
+        }
+
+        private void SpawnCloakedChest(SceneDirector sceneDirector)
+        {
             if (!SceneInfo.instance.countsAsStage && !SceneInfo.instance.sceneDef.allowItemsToSpawnObjects)
             {
                 return;
             }
 
-            for (int j = 0; j < chestCount; j++)
+            for (int j = 0; j < chestCount.Value; j++)
             {
-                Xoroshiro128Plus xoroshiro128Plus = new Xoroshiro128Plus(self.rng.nextUlong);
+                Xoroshiro128Plus xoroshiro128Plus = new Xoroshiro128Plus(sceneDirector.rng.nextUlong);
                 DirectorCore.instance.TrySpawnObject(new DirectorSpawnRequest(iscChest2Stealthed, new DirectorPlacementRule
                 {
                     placementMode = DirectorPlacementRule.PlacementMode.Random
